Update and rename TigerShark.py to tiger_shark.py
added module docstring
